complete backend

+++ backend/controller/goalController.js
@@ -1,6 +1,7 @@
 const asyncHandler = require('express-async-handler');
 
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel');
 
 /**
  * @desc  Get goals
@@ -8,7 +9,8 @@ const Goal = require('../models/goalModel');
  * @access Private
  */
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  //req.user.id - set in the protect middleware
+  const goals = await Goal.find({ user: req.user.id });
 
   res.status(200).json(goals);
 });
@@ -26,6 +28,7 @@ const setGoal = asyncHandler(async (req, res) => {
 
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
 
   res.status(200).json({ message: 'Goal Created', goal });
@@ -42,6 +45,21 @@ const updateGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error('Goal not found');
+  }
+
+  //ensure only owner of the goal can update it
+  const user = await User.findById(req.user.id);
+
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error('User not found');
+  }
+
+  //make sure logged in user matches goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error('User not authorized');
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -64,6 +82,21 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error('Goal not found');
+  }
+
+  //ensure only owner of the goal can delete it
+  const user = await User.findById(req.user.id);
+
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error('User not found');
+  }
+
+  //make sure logged in user matches goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error('User not authorized');
   }
 
   await Goal.findByIdAndDelete(req.params.id);
